+++ example/lib/views/inputs/inputs.dart
@@ -1,16 +1,16 @@
 library;
 
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:layrz_icons/layrz_icons.dart';
 import 'package:layrz_theme/layrz_theme.dart';
 import 'package:layrz_theme_example/store/store.dart';
 
 part 'src/buttons.dart';
-part 'src/text.dart';
-part 'src/checkboxes.dart';
-part 'src/radio.dart';
 part 'src/calendar.dart';
+part 'src/checkboxes.dart';
 part 'src/code.dart';
-
+part 'src/radio.dart';
 part 'src/selectors/datetime.dart';
 part 'src/selectors/general.dart';
+part 'src/text.dart';